+++ src/collective/smartlink/__init__.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 """Init and utils."""
-import logging
 from zope.i18nmessageid import MessageFactory
 
-_ = MessageFactory('collective.smartlink')
+import logging
+
+
+_ = MessageFactory("collective.smartlink")
 logger = logging.getLogger(__name__)