+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 1.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Plone 6 compatibility. Use 1.x tags for Plone < 6. [cekk]
+  [cekk]
 
 1.1.0 (2020-03-06)
 ------------------